gallium/hud: add frametime graph
Frametimes are often a more useful metric than FPS, because fps are an
averaged-value for the duration of a rather long period (one second).

Example: 59 frames are rendered in 16 ms each, but one frame takes 56 ms
to compute. That's exactly 60 FPS, but there is very noticeable stutter.
Some modern game engines and game testers also utilise frametimes graphs.

This graph is enabled via GALLIUM_HUD=frametime and will show the
frametime of each frame and therefore also the stutter in a
comprehensible way. It is useful to use the ceiling and/or dynamic
modifier, though, to clip extreme anomalies, e.g. during loading.